requirements.txt
google-genai~=1.41.0
openai>=1.0.0
Jinja2~=3.1.6

# Testing
pytest~=8.4.2
pytest-cov~=7.0.0
pytest-json-report~=1.5.0
